chore(clean codebase): - refactor codebase to comply with airbnb style guide [Finishes #156882304]
