+++ .eslintrc.json
@@ -4,6 +4,8 @@
 	"rules": {
 		"no-param-reassign": "off",
 		"no-case-declarations": "off",
+		"import/extensions": "off",
+    "import/no-unresolved": "off",
 		"jsx-a11y/label-has-for": [0],
 		"jsx-a11y/anchor-is-valid": "off",
 		"jsx-a11y/click-events-have-key-events": "off",
