+++ client/src/app/components/Sign/SigninComponent.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import OtherActions from '../../actions/others';
+import OtherActions from '../../actions/otherActions';
 
 /**
  * SignIn component class
