+++ client/src/app/components/Sign/SignupComponent.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import OtherActions from '../../actions/others';
+import OtherActions from '../../actions/otherActions';
 
 /**
  * SignIn component class
